requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
filelock==3.25.2
flatbuffers==25.12.19
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
idna==3.11
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.3
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PySocks==1.7.1
python-multipart==0.0.22
requests==2.32.5
rich==14.3.3
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
tensorflow==2.21.0
termcolor==3.3.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
wrapt==2.1.2
bcrypt==4.0.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
passlib==1.7.4
python-jose==3.3.0
sqlalchemy==2.0.41
email-validator==2.2.0
